+++ geopandas/_vectorized.py
@@ -8,15 +8,13 @@ import warnings
 
 import numpy as np
 import pandas as pd
-
 import shapely
 import shapely.geometry
 import shapely.geos
 import shapely.ops
+import shapely.validation
 import shapely.wkb
 import shapely.wkt
-import shapely.validation
-
 from shapely.geometry.base import BaseGeometry
 
 from . import _compat as compat
@@ -731,6 +729,25 @@ def interiors(data):
     return data
 
 
+def remove_repeated_points(data, tolerance=0.0):
+    if compat.USE_SHAPELY_20:
+        return shapely.remove_repeated_points(data, tolerance=tolerance)
+    if compat.USE_PYGEOS and compat.SHAPELY_GE_20:
+        warnings.warn(
+            "PyGEOS does not support remove_repeated_points, and Shapely >= 2 is "
+            "installed, thus using Shapely and not PyGEOS to remove repeated points.",
+            stacklevel=4,
+        )
+        return pygeos.from_shapely(
+            shapely.remove_repeated_points(to_shapely(data), tolerance=tolerance)
+        )
+    else:
+        raise NotImplementedError(
+            f"shapely >= 2.0 is required, "
+            f"version {shapely.__version__} is installed"
+        )
+
+
 def representative_point(data):
     if compat.USE_PYGEOS:
         return pygeos.point_on_surface(data)
@@ -1093,9 +1110,10 @@ def _shapely_normalize(geom):
     """
     Small helper function for now because it is not yet available in Shapely.
     """
-    from shapely.geos import lgeos
+    from ctypes import c_int, c_void_p
+
     from shapely.geometry.base import geom_factory
-    from ctypes import c_void_p, c_int
+    from shapely.geos import lgeos
 
     lgeos._lgeos.GEOSNormalize_r.restype = c_int
     lgeos._lgeos.GEOSNormalize_r.argtypes = [c_void_p, c_void_p]